# src/main/java/davincif/the_budget_project/login/request/LoginRequest.java
package davincif.the_budget_project.login.request;

import lombok.Data;
import lombok.experimental.Accessors;

@Data
@Accessors(chain = true)
public class LoginRequest {

    private String email;
    private String password;
}